First Commit: Created .gitignore and initial code files
This commit adds the .gitignore file to the repository to exclude the venv/
directory and .DS_Store files from version control. It also includes the
initial code files, Recieve.ino and controlCommands.cpp, which contain the
basic setup and functions for controlling a motor using the Adafruit MCP2515
FeatherWing CAN Receiver.